refactor(test): Make test filename consistent
