+++ src/macros/mod.rs
@@ -2,4 +2,4 @@ mod dispatch;
 mod seq;
 
 #[cfg(test)]
-mod test;
+mod tests;
